split out list module
