+++ src/pages/profile/index.tsx
@@ -1,6 +1,7 @@
 import { Layout } from "@/components/layout";
 import { HeaderLink } from "@/components/layout/header";
-import { Card, InfiniteScroll, NoResults } from "@/components/List";
+import { InfiniteScroll } from "@/components/list/infinite-scroll";
+import { Card, NoResults } from "@/components/list/product";
 import { Spinner } from "@/components/Loading";
 import { HeaderSearchBar, SEARCH_QUERY_KEY, SortDialog, useParseSort } from "@/components/Search";
 import { Star, UserPic } from "@/components/UI";
